% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_replication.R
\name{create_replication}
\alias{create_replication}
\title{Create replication object}
\usage{
create_replication(description_list, packages = NULL, project_path = NULL,
  data_list, function_script_path, replication_script_path, quietly = FALSE,
  checks = TRUE)
}
\arguments{
\item{description_list}{DESCRIPTION.}

\item{packages}{DESCRIPTION.}

\item{project_path}{DESCRIPTION.}

\item{data_list}{DESCRIPTION.}

\item{function_script_path}{DESCRIPTION.}

\item{replication_script_path}{DESCRIPTION.}

\item{quietly}{DESCRIPTION.}

\item{checks}{DESCRIPTION.}
}
\value{
RETURN DESCRIPTION
}
\description{
FUNCTION DESCRIPTION
}
\examples{
# ADD EXAMPLES HERE

}

