% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ipak}
\alias{ipak}
\title{FUNCTION TITLE}
\usage{
ipak(pkg, quietly = FALSE)
}
\arguments{
\item{pkg}{DESCRIPTION.}

\item{quietly}{DESCRIPTION.}
}
\value{
RETURN DESCRIPTION
}
\description{
FUNCTION DESCRIPTION
}
\examples{
# ADD EXAMPLES HERE

}

