% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replicate-deprecated.r
\name{Replicate-deprecated}
\alias{Replicate-deprecated}
\title{Deprecated functions in package \pkg{Replicate}.}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. Alternative functions with improved
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("<function>-deprecated")}.
}
\keyword{internal}
