% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replicate.R
\name{stronger_than-deprecated}
\alias{stronger_than-deprecated}
\alias{stronger_than}
\title{Probability of true effect stronger than threshold of scientific importance}
\usage{
stronger_than(q, yr, vyr = NULL, t2, vt2 = NULL, CI.level = 0.95,
  tail)
}
\arguments{
\item{q}{True effect size that is the threshold for "scientific importance"}

\item{yr}{Average true effect estimated using replications}

\item{vyr}{Estimated variance of above estimate}

\item{t2}{Heterogeneity of true effects estimated using replications}

\item{vt2}{Estimated variance of above estimate}

\item{CI.level}{Confidence level as a proportion}

\item{tail}{\code{above} for the probability of an effect above \code{q}; \code{below} for
the probability of an effect below \code{q}.}
}
\description{
This function is now deprecated. You should use \code{MetaUtility::prop_stronger} instead, which 
provides better, more general functionality. The below documentation is temporarily maintained before 
the \code{stronger_than} function is deprecated.
}
\details{
Given the original study's effect estimate and its variance, the estimated average true effect size in the 
replications, and the estimated heterogeneity in the replications, computes estimated probability that 
the original study would have an effect estimate at least as extreme as the observed value if the original
and the replications in fact are statistically consistent. Allows for heterogeneity.
}
\seealso{
\code{\link{Replicate-deprecated}}
}
\keyword{internal}
