% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.replication}
\alias{summary.replication}
\title{Summary method for replication class}
\usage{
\method{summary}{replication}(object, table = NULL, reported = FALSE,
  registered = FALSE, script = FALSE, desc = NULL, ...)
}
\arguments{
\item{object}{DESCRIPTION.}

\item{table}{DESCRIPTION.}

\item{reported}{DESCRIPTION.}

\item{registered}{DESCRIPTION.}

\item{script}{DESCRIPTION.}

\item{desc}{DESCRIPTION.}

\item{...}{DESCRIPTION.}
}
\value{
RETURN DESCRIPTION
}
\description{
FUNCTION DESCRIPTION
}
\examples{
# ADD EXAMPLES HERE

}

