% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_replication.R
\name{create_replication}
\alias{create_replication}
\title{Create replication object}
\usage{
create_replication(description_list, packages = NULL, project_path = NULL,
  data_list, function_script_path, replication_script_path, quietly = FALSE,
  checks = TRUE)
}
\arguments{
\item{description_list}{List of miscellaneous descriptions of replication.}

\item{packages}{Character vector of packages required for replication.}

\item{project_path}{Character string giving the path to the directory, where function and replication scripts are stored. Defaults to \code{NULL}, which is reasonable if RStudio project is used.}

\item{data_list}{A named list of data-frames used for the replication.}

\item{function_script_path}{Character string giving the name of \code{R} script which contains all functions required for the replication. The script should only include declarations of named functions using \code{<-} operator.}

\item{replication_script_path}{Character string giving the name of \code{R} script which contains all calls for table replications. The script should consist of only single calls for replication of one table and the table objects should be created using \code{<-} operator.}

\item{quietly}{Logical. Whether the creation of replication should go without any messages printed to console.}

\item{checks}{Logical. If \code{quietly = FALSE}, whether the checks for packages and consistency of replication should be performed.}
}
\value{
Replication class object
}
\description{
he function takes main parts of replication object as an arguments and returns the replication class object.
}
\examples{
# To be written...

}

