% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.replication}
\alias{summary.replication}
\title{Summary method for replication class}
\usage{
\method{summary}{replication}(object, table = NULL, reported = FALSE,
  registered = FALSE, script = FALSE, desc = NULL, ...)
}
\arguments{
\item{object}{Object of class replication created by \code{create_replication()}.}

\item{table}{Character string specifying the table to be replicated. The table name should include number of the table as specified in \code{replication_script_path} file and the word "table".}

\item{reported}{Logical. Whether to show columns with specifications reported in the paper.}

\item{registered}{Logical. Whether to show columns with specifications registered in PAP.}

\item{script}{Logical. Whether to print the script to replicate the results of the study. If \code{table = NULL}, then returns preamble which includes all the functions and packages required for replication. If \code{table != NULL}, then returns preamble and the code for replication of the specified table.}

\item{desc}{To be implemented...}

\item{...}{To be implemented...}
}
\value{
RETURN DESCRIPTION
}
\description{
FUNCTION DESCRIPTION
}
\examples{
# ADD EXAMPLES HERE

}

